msm8998-common: rootdir: Load WLAN driver during init.rc
'wificond' service gets started late in the stage of boot which looks for
'/sys/module/wlan/parameters/fwpath' but this module param is not created
until the WiFi driver is loaded.

'wificond' service is written assuming driver is already loaded
(statically built driver). Now to make WiFi turn on from GUI,
we need to mimic the driver as if it is statically built,
in other words WiFi driver should be loaded before starting the
wificond service from init.rc script.

With this change, WLAN driver is always loaded and never unloads.

Change-Id: I8fc05f3187bb03906d6db775cf634173c3248b58
CRs-Fixed: 1116914
Signed-off-by: Arian <arian.kulmer@web.de>